Expose zset-max-ziplist-value config via FP
We've received a use-case request where the user would like to have
the zset-max-ziplist-value configurable as the minimum size of an
item's data is larger than the default values per item in the zset
ziplist representation. This would convert all zsets to non-ziplist
representations which have a higher memory overhead.

This patch allows setting this via FP to make more efficient use
of memory, i.e. reduce memory overhead for zsets in the cluster.